Colorful Playground

+++ TheDiplomWork/OtherStuff/GeneralProgrammingStuff.cs
@@ -99,8 +99,8 @@ namespace TheDiplomWork
 
         public class Point3D
         {
-            public double x, y, z;
-            public Point3D(double _x, double _y, double _z)
+            public float x, y, z;
+            public Point3D(float _x, float _y, float _z)
             {
                 x = _x; y = _y; z = _z;
             }
